+++ src/components/ReaderLayout/ReaderLayout.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from "react";
 import * as DS from "@nypl/design-system-react-components";
 import { breadcrumbTitles } from "~/src/constants/labels";
 import { ApiLink, LinkResult } from "~/src/types/LinkQuery";
-import { MediaTypes } from "~/src/constants/mediaTypes";
 import IFrameReader from "../IFrameReader/IFrameReader";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import Layout from "../Layout/Layout";
@@ -11,6 +10,7 @@ import { formatUrl, truncateStringOnWhitespace } from "~/src/util/Util";
 import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/router";
+import { MediaTypes } from "~/src/constants/mediaTypes";
 const WebReader = dynamic(() => import("@nypl/web-reader"), { ssr: false });
 //The NYPL wrapper that wraps the Reader pages.
 const ReaderLayout: React.FC<{ linkResult: LinkResult }> = (props) => {
@@ -23,6 +23,11 @@ const ReaderLayout: React.FC<{ linkResult: LinkResult }> = (props) => {
 
   const isEmbed = MediaTypes.embed.includes(link.media_type);
   const isRead = MediaTypes.read.includes(link.media_type);
+
+  useEffect(() => {
+    gtag.drbEvents("Read", `${link.work.title}`);
+  }, [link]);
+
   return (
     <>
       {isEmbed && (