+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Added EDD button to the edition page
-- Added strict nodejs and npm version for deveopment
+- Added strict nodejs and npm version for development
 
 ## [0.10.4]